refactor: :recycle: Add `utils.rs` file to move utility functions away from `handler.rs`
